make constellation an optional dependency
