+++ amadeus-aws/README.md
@@ -1,0 +1,3 @@
+# amadeus-aws
+
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project includes a filesystem backend for S3 and a source for AWS Cloudfront logs.
