+++ amadeus-commoncrawl/README.md
@@ -1,0 +1,3 @@
+# amadeus-commoncrawl
+
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project includes a source for the CommonCrawl datasets.
