+++ amadeus-core/README.md
@@ -1,0 +1,3 @@
+# amadeus-core
+
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project includes fundamental definitions including `DistributedIterator` and `ProcessPool`.
