+++ amadeus-derive/README.md
@@ -1,8 +1,3 @@
 # amadeus-derive
 
-A crate for automatically deriving `amadeus::Data` for arbitrary structs.
-
-This crate is intended to be used by the `amadeus` crate. See [its documentation](https://docs.rs/amadeus) for usage.
-
-## License
-Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0.
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project provides the procedural macro `#[derive(Data)]` to automatically derive `amadeus::Data`.
