+++ amadeus-parquet/README.md
@@ -1,0 +1,3 @@
+# amadeus-parquet
+
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project includes a source for Parquet files.
