--- amadeus-parquet/amadeus-parquet-derive/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "amadeus-parquet-derive"
-version = "0.1.1"
-license = "Apache-2.0"
-authors = ["Alec Mocatta <alec@mocatta.net>"]
-categories = ["concurrency", "science", "database", "parser-implementations", "text-processing"]
-keywords = ["amadeus", "data", "parquet", "arrow", "hadoop"]
-description = """
-Apache Parquet Derive Macro implementation.
-"""
-repository = "https://github.com/alecmocatta/amadeus"
-homepage = "https://github.com/alecmocatta/amadeus"
-documentation = "https://docs.rs/amadeus/0.1.1"
-readme = "README.md"
-edition = "2018"
-
-[badges]
-azure-devops = { project = "alecmocatta/amadeus", pipeline = "tests" }
-maintenance = { status = "actively-developed" }
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = { version = "1.0.1", default-features = false }
-quote = { version = "1.0.2", default-features = false }
-syn = { version = "1.0.5", default-features = false, features = ["derive", "clone-impls", "parsing", "printing", "proc-macro"] }
