--- amadeus-parquet/amadeus-parquet-derive/README.md
@@ -1,8 +1,0 @@
-# amadeus-parquet-derive
-
-A crate for automatically deriving `amadeus_parquet::record::Record` for arbitrary structs.
-
-This crate is intended to be used by the `amadeus-parquet` crate. See [its documentation](https://docs.rs/amadeus-parquet) for usage.
-
-## License
-Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0.
