+++ amadeus-postgres/README.md
@@ -1,0 +1,3 @@
+# amadeus-postgres
+
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project includes a source for PostgreSQL databases.
