+++ amadeus-serde/README.md
@@ -1,0 +1,3 @@
+# amadeus-serde
+
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project includes source for JSON and CSV files.
