+++ amadeus-types/README.md
@@ -1,0 +1,3 @@
+# amadeus-types
+
+This subcrate of the [`amadeus`](https://github.com/alecmocatta/amadeus) project includes common types such as `Decimal`, `Timestamp` and `Url`.
