+++ src/source.rs
@@ -5,14 +5,14 @@ pub use amadeus_aws::Cloudfront;
 pub mod aws {
 	pub use crate::data::CloudfrontRow;
 	#[doc(inline)]
-	pub use amadeus_aws::*;
+	pub use amadeus_aws::{AwsError, AwsRegion, S3Directory, S3File};
 }
 #[cfg(feature = "commoncrawl")]
 #[doc(inline)]
 pub use amadeus_commoncrawl::{self as commoncrawl, CommonCrawl};
 #[cfg(feature = "parquet")]
 #[doc(inline)]
-pub use amadeus_parquet::{self as parquet, Parquet};
+pub use amadeus_parquet::{self as parquet, Parquet, ParquetDirectory};
 #[cfg(feature = "postgres")]
 #[doc(inline)]
 pub use amadeus_postgres::{self as postgres, Postgres};
