+++ tests/cloudfront.rs
@@ -3,15 +3,15 @@
 #[macro_use]
 extern crate serde_closure;
 
-use amadeus::{
-	prelude::*, source::aws::{CloudfrontRow, Error}
-};
+use amadeus::prelude::*;
+#[cfg(feature = "constellation")]
 use constellation::*;
 use std::{
 	env, time::{Duration, SystemTime}
 };
 
 fn main() {
+	#[cfg(feature = "constellation")]
 	init(Resources::default());
 
 	// Accept the number of processes at the command line, defaulting to 10
@@ -28,10 +28,13 @@ fn main() {
 		let thread_pool = ThreadPool::new(processes).unwrap();
 		run(&thread_pool)
 	};
+	#[cfg(feature = "constellation")]
 	let process_pool_time = {
 		let process_pool = ProcessPool::new(processes, 1, Resources::default()).unwrap();
 		run(&process_pool)
 	};
+	#[cfg(not(feature = "constellation"))]
+	let process_pool_time = "-";
 
 	println!(
 		"in {:?} {:?} {:?}",
@@ -42,7 +45,7 @@ fn main() {
 fn run<P: amadeus_core::pool::ProcessPool>(pool: &P) -> Duration {
 	let start = SystemTime::now();
 
-	let _ = DistributedIteratorMulti::<&Result<CloudfrontRow, Error>>::count(Identity);
+	let _ = DistributedIteratorMulti::<&Result<CloudfrontRow, AwsError>>::count(Identity);
 
 	let ((), (count, count2)) = Cloudfront::new(
 		rusoto_core::Region::UsEast1,
@@ -60,7 +63,7 @@ fn run<P: amadeus_core::pool::ProcessPool>(pool: &P) -> Duration {
 		(
 			Identity.map(FnMut!(|_x: &Result<_, _>| {})).count(),
 			Identity.cloned().count(),
-			// DistributedIteratorMulti::<&Result<CloudfrontRow, Error>>::count(Identity),
+			// DistributedIteratorMulti::<&Result<CloudfrontRow, AwsError>>::count(Identity),
 		),
 	);
 	assert_eq!(count, count2);
